==> ML_Final/1_DatasetFromParquet.ipynb <==
# путь до данных на компьютере
path = 'train_data/'

import os
import pandas as pd
import tqdm


def read_parquet_dataset_from_local(path_to_dataset: str, start_from: int = 0,
                                     num_parts_to_read: int = 2, columns=None, verbose=False) -> pd.DataFrame:
    """
    читает num_parts_to_read партиций, преобразовывает их к pd.DataFrame и возвращает
    :param path_to_dataset: путь до директории с партициями
    :param start_from: номер партиции, с которой нужно начать чтение
    :param num_parts_to_read: количество партиций, которые требуется прочитать
    :param columns: список колонок, которые нужно прочитать из партиции
    :return: pd.DataFrame
    """

    res = []
    dataset_paths = sorted([os.path.join(path_to_dataset, filename) for filename in os.listdir(path_to_dataset)
                              if filename.startswith('train')])
    print(dataset_paths)

    start_from = max(0, start_from)
    chunks = dataset_paths[start_from: start_from + num_parts_to_read]
    if verbose:
        print('Reading chunks:\n')
        for chunk in chunks:
            print(chunk)
    for chunk_path in tqdm.tqdm_notebook(chunks, desc="Reading dataset with pandas"):
        print('chunk_path', chunk_path)
        chunk = pd.read_parquet(chunk_path,columns=columns)
        res.append(chunk)

    return pd.concat(res).reset_index(drop=True)

def prepare_transactions_dataset(path_to_dataset: str, num_parts_to_preprocess_at_once: int = 1, num_parts_total: int=50,
                                 save_to_path=None, verbose: bool=False):
    """
    возвращает готовый pd.DataFrame с признаками, на которых можно учить модель для целевой задачи
    path_to_dataset: str
        путь до датасета с партициями
    num_parts_to_preprocess_at_once: int
        количество партиций, которые будут одновременно держаться и обрабатываться в памяти
    num_parts_total: int
        общее количество партиций, которые нужно обработать
    save_to_path: str
        путь до папки, в которой будет сохранён каждый обработанный блок в .parquet-формате; если None, то не будет сохранён
    verbose: bool
        логирует каждую обрабатываемую часть данных
    """
    preprocessed_frames = []

    for step in tqdm.tqdm_notebook(range(0, num_parts_total, num_parts_to_preprocess_at_once),
                                   desc="Transforming transactions data"):
        transactions_frame = read_parquet_dataset_from_local(path_to_dataset, step, num_parts_to_preprocess_at_once,
                                                             verbose=verbose)


   #здесь должен быть препроцессинг данных
    
    
   #записываем подготовленные данные в файл
        if save_to_path:
            block_as_str = str(step)
            if len(block_as_str) == 1:
                block_as_str = '00' + block_as_str
            else:
                block_as_str = '0' + block_as_str
            transactions_frame.to_parquet(os.path.join(save_to_path, f'processed_chunk_{block_as_str}.parquet'))

        preprocessed_frames.append(transactions_frame)
    return pd.concat(preprocessed_frames)

data = prepare_transactions_dataset(path, num_parts_to_preprocess_at_once=12, num_parts_total=12) 

data.to_csv('dataset/dataset.csv', index = False)

import pandas as pd
df = pd.read_csv('dataset/dataset.csv')

# Преобразуем все колонки, кроме id , в int8
df_id = df.id
df=df.drop('id', axis = 1)
df=df.astype('int8')
df['id'] = df_id
df.head()

df.to_csv('dataset/dataset_int8.csv', index = False)

#сделаем  небольшую выборку для разведочного анализа
import pandas as pd
df_target = pd.read_csv('train_data/train_target.csv')
df_target.shape

df_target_sample = df_target.sample(n=30000, random_state = 1)

# Посмотрим на соотношение значений таргета
print(f'Исходный df_target: {df_target.flag.value_counts()}')
print(f'df_target_sample: {df_target_sample.flag.value_counts()}')

df_target_sample = df_target_sample.set_index('id')
df_target_sample.head()

# соберем датасет со всеми строками,относящимися к id, попавшими в семпл-выборку 
#(пока не аггрегируем -хочу иметь возможность попробовать на семпле разную аггрегацию)
df_joined = df.join(df_target_sample, on = 'id', how = 'inner')
df_joined.head()

df_joined.shape

df_joined.id.nunique()

df_joined.to_csv('dataset/dataset_sample.csv', index = False)
